+++ openfisca_core/periods.py
@@ -123,8 +123,8 @@ class Instant(tuple):
         >>> instant('2014-2-3').period('day', size = 2)
         Period((u'day', Instant((2014, 2, 3)), 2))
         """
-        assert unit in (u'day', u'month', u'year'), 'Invalid unit: {} of type {}'.format(unit, type(unit))
-        assert isinstance(size, int) and size >= 1, 'Invalid size: {} of type {}'.format(size, type(size))
+        assert unit in (u'day', u'month', u'year'), u'Invalid unit: {} of type {}'.format(unit, type(unit))
+        assert isinstance(size, int) and size >= 1, u'Invalid size: {} of type {}'.format(size, type(size))
         return Period((unicode(unit), self, size))
 
     def offset(self, offset, unit):
@@ -212,18 +212,18 @@ class Instant(tuple):
             if unit == u'month':
                 day = 1
             else:
-                assert unit == u'year', 'Invalid unit: {} of type {}'.format(unit, type(unit))
+                assert unit == u'year', u'Invalid unit: {} of type {}'.format(unit, type(unit))
                 month = 1
                 day = 1
         elif offset == 'last-of':
             if unit == u'month':
                 day = calendar.monthrange(year, month)[1]
             else:
-                assert unit == u'year', 'Invalid unit: {} of type {}'.format(unit, type(unit))
+                assert unit == u'year', u'Invalid unit: {} of type {}'.format(unit, type(unit))
                 month = 12
                 day = 31
         else:
-            assert isinstance(offset, int), 'Invalid offset: {} of type {}'.format(offset, type(offset))
+            assert isinstance(offset, int), u'Invalid offset: {} of type {}'.format(offset, type(offset))
             if unit == u'day':
                 day += offset
                 if offset < 0:
@@ -256,7 +256,7 @@ class Instant(tuple):
                 if day > month_last_day:
                     day = month_last_day
             else:
-                assert unit == u'year', 'Invalid unit: {} of type {}'.format(unit, type(unit))
+                assert unit == u'year', u'Invalid unit: {} of type {}'.format(unit, type(unit))
                 year += offset
                 # Handle february month of leap year.
                 month_last_day = calendar.monthrange(year, month)[1]
@@ -343,7 +343,7 @@ class Period(tuple):
 
     @property
     def date(self):
-        assert self.size == 1, '"date" is undefined for a period of size > 1: {}'.format(self)
+        assert self.size == 1, u'"date" is undefined for a period of size > 1: {}'.format(self)
         return self.start.date
 
     @property
@@ -622,7 +622,7 @@ class Period(tuple):
                     year += 1
                     month -= 12
             else:
-                assert unit == u'year', 'Invalid unit: {} of type {}'.format(unit, type(unit))
+                assert unit == u'year', u'Invalid unit: {} of type {}'.format(unit, type(unit))
                 year += size
             day -= 1
             if day < 1: